streamlit/requirements.txt
agate==1.7.1
altair==5.1.2
attrs==23.1.0
babel==2.13.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
db-dtypes==1.1.1
dbt-bigquery==1.6.7
dbt-core==1.6.6
dbt-extractor==0.4.1
dbt-semantic-interfaces==0.2.2
fonttools==4.43.1
gitdb==4.0.11
gitpython==3.1.40
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.12.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-dataproc==5.6.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
hologram==0.0.16
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
leather==0.3.4
logbook==1.5.3
markdown-it-py==3.0.0
markupsafe==2.1.3
mashumaro==3.8.1
matplotlib==3.8.0
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.0.7
networkx==3.2
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
pandas-gbq==0.19.2
parsedatetime==2.6
pathspec==0.11.2
pillow==10.1.0
plotly==5.17.0
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydata-google-auth==1.8.2
pydeck==0.8.1b0
pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
streamlit==1.27.2
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
python-dotenv